[WIP] Implement OpenLab CDS (.dx) Part 2
This commit implements actual file reading for the .uv files in the
zipped .dx format used by Agilent OpenLab CDS software. It also includes
updates to the metadata extraction process to accommodate the structure
of OpenLab CDS files.

A new 'format' field is added to the metadata to differentiate between
Chemstation and OpenLab CDS files. The commit also includes updates to
the test output files to reflect this new format field.